Separate basic elasto items and components
