+++ include/seahowl/aero/aerodyn_adapter.h
@@ -6,7 +6,7 @@
 #include <seahowl/core/turbine.h>
 #include <seahowl/elasto/blade_elasto.h>
 #include <seahowl/elasto/rotor_elasto.h>
-#include <seahowl/elasto/elasto.h>
+#include <seahowl/elasto/entities_elasto.h>
 #include <seahowl/commons.h>
 
 /// <summary>
