+++ include/seahowl/elasto/chrono_adapters.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <seahowl/elasto/reference_point_elasto.h>
-#include <seahowl/elasto/elasto.h>
+#include <seahowl/elasto/entities_elasto.h>
 
 #include <vector>
 #include <memory>
