+++ include/seahowl/elasto/rotor_elasto.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <seahowl/elasto/elasto.h>
+#include <seahowl/elasto/component_elasto.h>
 
 #include <vector>
 #include <memory>
