+++ include/seahowl/io/write_vtk.h
@@ -1,4 +1,4 @@
-#include <seahowl/elasto/elasto.h>
+#include <seahowl/elasto/component_elasto.h>
 
 // Forward declaration
 class vtkUnstructuredGrid;
