+++ src/core/blade.cpp
@@ -2,7 +2,6 @@
 
 #include <seahowl/elasto/reference_point_elasto.h>
 #include <seahowl/elasto/blade_elasto.h>
-#include <seahowl/elasto/elasto.h>
 #include <seahowl/aero/blade_aero.h>
 #include <seahowl/core/utils.h>
