+++ src/elasto/component_elasto.cpp
@@ -1,4 +1,4 @@
-#include <seahowl/elasto/elasto.h>
+#include <seahowl/elasto/component_elasto.h>
 #include <seahowl/elasto/chrono_adapters.h>
 
 #include <vector>
